// guara/src/types/apiTypes.ts
export interface LoginResponse {
  token: string;
  user: string; // username
  email: string;
  permissao: string;
  repositorios_associados_nomes: string[];
  validade: string; // Formato ISO de data
}

// Representa os dados do usuário logado, para serem armazenados no Pinia store
export interface AuthenticatedUser {
  username: string;
  email: string;
  permission: string;
  associatedRepositories: string[];
}